Luanching the application on emulator

+++ app/src/main/java/com/example/mychannels/Model/BookRecording.kt
@@ -3,5 +3,5 @@ package com.example.mychannels.Model
 data class BookRecording(
     val recordingId: Int,
     val title :String,
-    val duration: Int
+    val duration: Long
 )
